fix(pipectl): retry transient client errors correctly
